use_generated_srt parameter added

+++ main.py
@@ -25,6 +25,7 @@ def process_video_with_subtitles(
     marginv: int = 10,  # Vertical margin (pixels)
     model_size: str = "large",  # tiny, base, small, medium, large (large-v2, large-v3)
     keep_srt: bool = True,
+    use_genererated_srt: bool = False,
 ):
     """
     Complete pipeline: transcribe audio and embed subtitles into video.
@@ -40,32 +41,39 @@ def process_video_with_subtitles(
     if folder_path and not os.path.exists(folder_path):
         os.makedirs(folder_path)
 
+    flag = True
     # 1. Generate subtitles
-    if not generate_subtitles(input_video, temp_srt, model_size, language):
-        return
+    if use_genererated_srt:
+        if not os.path.exists(temp_srt):
+            flag = generate_subtitles(
+                input_video, temp_srt, model_size, language
+            )
+    else:
+        flag = generate_subtitles(input_video, temp_srt, model_size, language)
 
-    subtitle_style = get_subtitle_style(
-        font=font,
-        fontsize=fontsize,
-        primary_colour=font_colour,
-        marginv=marginv,
-    )
+    if flag:
+        subtitle_style = get_subtitle_style(
+            font=font,
+            fontsize=fontsize,
+            primary_colour=font_colour,
+            marginv=marginv,
+        )
 
-    # 2. Add subtitles to video
-    success = add_subtitles_to_video(
-        input_video=input_video,
-        input_srt=temp_srt,
-        output_video=output_video,
-        hardcode=hardcode,
-        subtitle_style=subtitle_style,
-    )
+        # 2. Add subtitles to video
+        success = add_subtitles_to_video(
+            input_video=input_video,
+            input_srt=temp_srt,
+            output_video=output_video,
+            hardcode=hardcode,
+            subtitle_style=subtitle_style,
+        )
 
-    # 3. Cleanup
-    if success:
-        logger.info(f"✅ Success: Subtitled video saved to {output_video}")
-    if not keep_srt and os.path.exists(temp_srt):
-        os.remove(temp_srt)
-        logger.info(f"Temp subtitles file removed")
+        # 3. Cleanup
+        if success:
+            logger.info(f"✅ Success: Subtitled video saved to {output_video}")
+        if not keep_srt and os.path.exists(temp_srt):
+            os.remove(temp_srt)
+            logger.info(f"Temp subtitles file removed")
 
 
 if __name__ == "__main__":
@@ -78,4 +86,5 @@ if __name__ == "__main__":
         marginv=160,
         # model_size="small",
         # keep_srt=False,
+        # use_genererated_srt=True,
     )
